[Cocoa] Improve handling of instances, give a title to child windows, and fix destroy (#109)
* Handle instances & app termination in a better way

The active_instances property was added in d771625 to keep track of the
number of open instances, so that we could terminate the app when the
last window was closed. This is better acheived using the added
AppDelegate method.

Till now we could also not delete instances added to the `instances`
list once they were closed. This was probably because deleting them
released the associated Cocoa objects, and PyObjC's autorelease pool
tried to deallocate them again, resulting in a Segmentation Fault. This
is fixed by calling retain() on these objects on initialisation.

Hence now `instances` variable contains only open instances. This makes
active_instances obsolete and so it has been removed.

* Give a title to the child windows and fix destroy

* With the current design, there is no way to explicitly set the title
of the child windows (JS cannot set the title of the windows it opens).
Web browsers normally use the page title as the title for pop-up
windows, so we follow suit.

* Close instance on calling destroy, not quit app