Add window 'name' parameter with example
This parameter works as an 'instance id' for each pywebview window. It
is synonymous with JS window name property. Hence it can be used to
cross refer a window from both JS (as the target of links, forms) and
Python (in the native API).

Caveat: The name, once set upon creation with window.open, should not be
changed with JS, as this change cannot (currently) be updated in Python.

* Commits *

- Create skeleton for window name parameter
    Suggestion: make name a required parameter before merging into master.
Currently it is added as an optional parameter to create_window in the
init script, so that we do not break the current examples.

- [Cocoa] Add window name parameter

- [Cocoa] Add window name example